factor out types::Data
