+++ p2p/src/authenticated/lookup/actors/peer/actor.rs
@@ -1,10 +1,10 @@
 use super::{ingress::Message, Config, Error, Mailbox};
 use crate::authenticated::{
+    data::Data,
     lookup::{
         actors::tracker::{self, Reservation},
         channels::Channels,
-        metrics,
-        types::{self, Data},
+        metrics, types,
     },
     relay::Relay,
 };
@@ -29,8 +29,8 @@ pub struct Actor<E: Spawner + Clock + ReasonablyRealtime + Metrics, C: PublicKey
     ping_frequency: Duration,
     mailbox: Mailbox,
     control: mpsc::Receiver<Message>,
-    high: mpsc::Receiver<types::Data>,
-    low: mpsc::Receiver<types::Data>,
+    high: mpsc::Receiver<Data>,
+    low: mpsc::Receiver<Data>,
 
     sent_messages: Family<metrics::Message, Counter>,
     received_messages: Family<metrics::Message, Counter>,
@@ -66,10 +66,7 @@ impl<E: Spawner + Clock + ReasonablyRealtime + Rng + CryptoRng + Metrics, C: Pub
     }
 
     /// Unpack `msg` and verify the underlying `channel` is registered.
-    fn validate_msg<V>(
-        msg: Option<types::Data>,
-        rate_limits: &HashMap<u32, V>,
-    ) -> Result<types::Data, Error> {
+    fn validate_msg<V>(msg: Option<Data>, rate_limits: &HashMap<u32, V>) -> Result<Data, Error> {
         let data = match msg {
             Some(data) => data,
             None => return Err(Error::PeerDisconnected),
