+++ p2p/src/authenticated/mod.rs
@@ -1,3 +1,4 @@
+mod data;
 pub mod discovery;
 mod ip;
 pub mod lookup;
